2147 moving getPreviewComponentDefinition to a separate file
