+++ src/components/fields/fieldUtils.ts
@@ -16,7 +16,8 @@
  */
 
 import { Schema, SchemaDefinition } from "@/core";
-import { isExpression } from "@/runtime/mapArgs";
+import { isTemplateExpression } from "@/runtime/mapArgs";
+import { UnknownObject } from "@/types";
 
 export function fieldLabel(name: string): string {
   const parts = name.split(".");
@@ -50,10 +51,10 @@ export function createTypePredicate(predicate: TypePredicate): TypePredicate {
   };
 }
 
-export function getPreviewValues<TObj = any>(obj: TObj): TObj {
-  const newObj = {};
+export function getPreviewValues<TObj = UnknownObject>(obj: TObj): TObj {
+  const newObj: TObj = {} as any;
   for (const [key, value] of Object.entries(obj)) {
-    if (isExpression(value)) {
+    if (isTemplateExpression(value)) {
       newObj[key] = value.__value__;
     } else if (!Array.isArray(value) && typeof value === "object") {
       newObj[key] = getPreviewValues(value);
